Social Bar
I've added a social bar, but i can't get the alignment quite right without it looking silly. .bar is the social bar and it should move with the screen but it isnt. I've added social links to each icon "although he doesnt have a twitter yet" and ive linked the shop page to an external shop he has until we get things moving